Refactor pipelines to use common code
